requirements.txt
# Core Dependencies
groq>=0.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
asyncssh>=2.13.0
python-dotenv>=1.0.0

# API Server Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
psycopg2-binary>=2.9.0

# Security
cryptography>=41.0.0

# Scheduling
apscheduler>=3.10.0

# Utilities
httpx>=0.25.0
